requirements.txt
requests==2.24.0
jsonpath==0.82
pytest==5.4.3
allure-pytest==2.8.16
allure-python-commons==2.8.16
PyYAML==5.3.1
pytest-xdist==2.1.0
pytest-ordering==0.6